+++ opal/patients/tests/test_views.py
@@ -1,12 +1,38 @@
 from http import HTTPStatus
+from typing import Tuple
 
+from django.contrib.auth.models import AbstractUser
 from django.forms.models import model_to_dict
 from django.test import Client
 from django.urls import reverse
 
-from pytest_django.asserts import assertContains, assertQuerysetEqual
+import pytest
+from pytest_django.asserts import assertContains, assertQuerysetEqual, assertTemplateUsed
 
-from .. import factories, models, tables
+from .. import factories, forms, models, tables
+
+# Add any future GET-requestable patients app pages here for faster test writing
+test_url_template_data: list[Tuple] = [
+    (reverse('patients:relationships-search'), 'patients/relationships-search/form.html'),
+]
+
+
+@pytest.mark.parametrize(('url', 'template'), test_url_template_data)
+def test_patients_urls_exist(user_client: Client, admin_user: AbstractUser, url: str, template: str) -> None:
+    """Ensure that a page exists at each URL address."""
+    user_client.force_login(admin_user)
+    response = user_client.get(url)
+
+    assert response.status_code == HTTPStatus.OK
+
+
+@pytest.mark.parametrize(('url', 'template'), test_url_template_data)
+def test_views_use_correct_template(user_client: Client, admin_user: AbstractUser, url: str, template: str) -> None:
+    """Ensure that a page uses appropriate templates."""
+    user_client.force_login(admin_user)
+    response = user_client.get(url)
+
+    assertTemplateUsed(response, template)
 
 
 def test_relationshiptypes_list_table(user_client: Client) -> None:
@@ -154,3 +180,43 @@ def test_relationships_not_pending_not_list(user_client: Client) -> None:
     response = user_client.get(reverse('patients:relationships-pending-list'))
 
     assert len(response.context['relationship_list']) == 2
+
+
+def test_relationships_pending_list_table(user_client: Client) -> None:
+    """Ensures that pending relationships list uses the corresponding table."""
+    response = user_client.get(reverse('patients:relationships-pending-list'))
+
+    assert response.context['table'].__class__ == tables.PendingRelationshipTable
+
+
+def test_relationships_pending_form(user_client: Client) -> None:
+    """Ensures that pending relationships edit uses the right form."""
+    relationshiptype = factories.RelationshipType(name='relationshiptype')
+    factories.Relationship(pk=1, type=relationshiptype)
+    response = user_client.get(reverse('patients:relationships-pending-update', kwargs={'pk': 1}))
+
+    assert response.context['form'].__class__ == forms.RelationshipPendingAccessForm
+
+
+def test_relationships_pending_form_content(user_client: Client) -> None:
+    """Ensures that pending relationships passed info is correct."""
+    relationshiptype = factories.RelationshipType(name='relationshiptype')
+    caregiver = factories.CaregiverProfile()
+    relationship = factories.Relationship(pk=1, type=relationshiptype, caregiver=caregiver)
+    response = user_client.get(reverse('patients:relationships-pending-update', kwargs={'pk': 1}))
+
+    assert response.context['relationship'] == relationship
+
+
+def test_relationships_pending_form_response(user_client: Client) -> None:
+    """Ensures that pending relationships displayed info is correct."""
+    relationshiptype = factories.RelationshipType(name='relationshiptype')
+    caregiver = factories.CaregiverProfile()
+    patient = factories.Patient()
+    relationship = factories.Relationship(pk=1, type=relationshiptype, caregiver=caregiver, patient=patient)
+    response = user_client.get(reverse('patients:relationships-pending-update', kwargs={'pk': 1}))
+    response.content.decode('utf-8')
+
+    assertContains(response, patient)
+    assertContains(response, caregiver)
+    assertContains(response, relationship.patient.ramq)