+++ CHANGELOG.md
@@ -35,8 +35,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Add REST API endpoint for the patient's caregiver(s) relationship(s) (!222)
 * feat: Add can_answer_questionnaire field to RelationshipType model (!689)
 * Add unique constraint to Relationship model (!655)
+* Generate a random uuid for the caregiver username (!267)
+* Add a "Manage Caregiver Access to Patient Data" page (!271)
 * Add date_of_death field to Patients model (!714)
-* Generate a random uuid for the caregiver username(!267)
+* Generate a random uuid for the caregiver username (!267)
 
 ### Documentation